refactor(mail): combine encoder and decoder into codec crate
