+++ .github/workflows/ci.yml
@@ -91,7 +91,7 @@ jobs:
         run: cargo fmt --all -- --check
 
       - name: clippy
-        run: cargo clippy --workspace --all-targets -- -D warnings
+        run: cargo clippy --workspace --all-targets --all-features -- -D warnings
 
   test-rs:
     name: Test Rust (${{ matrix.os }})
@@ -123,4 +123,4 @@ jobs:
         run: cargo build --workspace --all-targets --verbose
 
       - name: Test
-        run: cargo test --workspace --verbose
+        run: cargo test --workspace --all-features --verbose
