+++ crates/apps/rokbattles-desktop/Cargo.toml
@@ -28,7 +28,7 @@ tauri-plugin-dialog = { workspace = true }
 anyhow = { workspace = true }
 tokio = { workspace = true, features = ["macros", "sync", "time"] }
 reqwest = { workspace = true, features = ["json", "multipart", "rustls-tls-native-roots"] }
-rokbattles-mail-decoder = { path = "../../common/rokbattles-mail-decoder" }
+rokbattles-mail-codec = { path = "../../common/rokbattles-mail-codec", features = ["read"] }
 rokbattles-mail-registry = { path = "../../common/rokbattles-mail-registry" }
 tauri-plugin-updater = { workspace = true }
 notify = { workspace = true }
