+++ crates/apps/rokbattles-processor/Cargo.toml
@@ -15,7 +15,7 @@ futures = { workspace = true, features = ["std"] }
 zstd = { workspace = true }
 sha2 = { workspace = true }
 dotenvy = { workspace = true }
-rokbattles-mail-decoder = { path = "../../common/rokbattles-mail-decoder" }
+rokbattles-mail-codec = { path = "../../common/rokbattles-mail-codec", features = ["read"] }
 rokbattles-mail-registry = { path = "../../common/rokbattles-mail-registry", features = ["processors"] }
 rokbattles-mail-sdk = { path = "../../common/rokbattles-mail-sdk" }
 sentry = { workspace = true, features = ["backtrace", "contexts", "debug-images", "panic", "reqwest", "rustls"] }
