--- crates/common/rokbattles-mail-decoder/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "rokbattles-mail-decoder"
-edition.workspace = true
-rust-version.workspace = true
-version.workspace = true
-
-[dependencies]
-rokbattles-djb2-simd = { path = "../rokbattles-djb2-simd" }
-serde_json = { workspace = true }
-thiserror = { workspace = true }
-
-[lib]
-doctest = false
-
-[lints]
-workspace = true
