--- crates/common/rokbattles-mail-encoder/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "rokbattles-mail-encoder"
-edition.workspace = true
-rust-version.workspace = true
-version.workspace = true
-
-[dependencies]
-rokbattles-djb2-simd = { path = "../rokbattles-djb2-simd" }
-serde_json = { workspace = true }
-thiserror = { workspace = true }
-
-[dev-dependencies]
-rokbattles-mail-decoder = { path = "../rokbattles-mail-decoder" }
-
-[lib]
-doctest = false
-
-[lints]
-workspace = true
